+++ btclib/block/__init__.py
@@ -7,9 +7,10 @@
 #
 # No part of btclib including this file, may be copied, modified, propagated,
 # or distributed except according to the terms contained in the LICENSE file.
-"""Module btclib.ecc."""
+"""Module btclib.block."""
 
+from btclib.block import merkle_proof
 from btclib.block.block import Block
 from btclib.block.block_header import BlockHeader
 
-__all__ = ["Block", "BlockHeader"]
+__all__ = ["Block", "BlockHeader", "merkle_proof"]